+++ packages/provider-utils/src/index.ts
@@ -14,7 +14,7 @@ export { DownloadError } from './download-error';
 export {
   readResponseWithSizeLimit,
   DEFAULT_MAX_DOWNLOAD_SIZE,
-} from './download-with-size-limit';
+} from './read-response-with-size-limit';
 export * from './fetch-function';
 export { createIdGenerator, generateId, type IdGenerator } from './generate-id';
 export * from './get-error-message';
