+++ packages/provider-utils/src/read-response-with-size-limit.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { readResponseWithSizeLimit } from './download-with-size-limit';
+import { readResponseWithSizeLimit } from './read-response-with-size-limit';
 import { DownloadError } from './download-error';
 
 function createMockResponse({
